backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0

# ─── Supabase ─────────────────────────────────────────────────────────────────
supabase>=2.9.0

# ─── Auth & Security ──────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0

# ─── AI ───────────────────────────────────────────────────────────────────────
google-genai>=1.0.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.24.0
